+++ crates/sncast/src/helpers/mod.rs
@@ -2,6 +2,7 @@ pub mod block_explorer;
 pub mod braavos;
 pub mod configuration;
 pub mod constants;
+pub mod deploy;
 pub mod error;
 pub mod fee;
 pub mod rpc;